Implemented JWT authentication login feature and tested.
